Andi's Updates
Because Andi Is working in VS Code He cant commit without messing with our eclipse so Im commitng for him to be more "Eclipse friendly". He added a lot of object serialzation. Changed the relationships between classes. Instead of using X's and O's we use Player Classes for Human Players and AI's.